+++ src/utils/scoreRender.ts
@@ -1,0 +1,4 @@
+export const scoreRender = (score: number) => {
+  const stars = ['⭐', '⭐⭐', '⭐⭐⭐', '⭐⭐⭐⭐', '⭐⭐⭐⭐⭐'];
+  return stars[score - 1] || null;
+};
